move signals into lib folder
