+++ apps/next/src/lib/signals.ts
@@ -1,5 +1,26 @@
-import { Signal, computed, effect } from '@preact/signals-core'
-import { useEffect, useMemo, useRef, useState } from 'react'
+// temporary solution from https://github.com/JonAbrams/signals-react-safe/tree/main
+
+import {} from 'react'
+import { type ReactElement, useEffect, useMemo, useRef, useState } from 'react'
+import {
+	Signal,
+	signal,
+	computed,
+	effect,
+	batch,
+	type ReadonlySignal,
+	untracked,
+} from '@preact/signals-core'
+
+export {
+	signal,
+	computed,
+	effect,
+	batch,
+	Signal,
+	type ReadonlySignal,
+	untracked,
+}
 
 export function useSignalValue<T>(signal: Signal<T>): T {
 	const [state, setState] = useState<T>(signal.value)
@@ -39,4 +60,26 @@ export function useSignalEffect(cb: () => void | (() => void)): void {
 	useEffect(() => {
 		return effect(() => callback.current())
 	}, [])
+}
+
+const ReactElemType = Symbol.for('react.element') // https://github.com/facebook/react/blob/346c7d4c43a0717302d446da9e7423a8e28d8996/packages/shared/ReactSymbols.js#L15
+
+function SignalValue({ data }: { data: Signal }) {
+	return useSignalValue(data)
+}
+
+Object.defineProperties(Signal.prototype, {
+	$$typeof: { configurable: true, value: ReactElemType },
+	type: { configurable: true, value: SignalValue },
+	props: {
+		configurable: true,
+		get() {
+			return { data: this }
+		},
+	},
+	ref: { configurable: true, value: null },
+})
+
+declare module '@preact/signals-core' {
+	interface Signal extends ReactElement {}
 }
